[claude-code] Fix PR review comments - refactor tool management
