+++ gigachat-ai-agent/src/main/java/ru/comavp/tools/adapters/ToolAdapter.java
@@ -1,4 +1,4 @@
-package ru.comavp.tools;
+package ru.comavp.tools.adapters;
 
 import chat.giga.model.completion.ChatFunction;
 import ru.comavp.tools.model.ToolResult;
@@ -8,7 +8,7 @@ import java.util.Map;
 /**
  * @author Claude Code
  */
-public interface UnifiedToolDefinition {
+public interface ToolAdapter {
 
     String getName();
     String getDescription();
